Public API

AdaptiveRejectionSampling.ARSampler

AdaptiveRejectionSampling.Objective

AdaptiveRejectionSampling.Objective

AdaptiveRejectionSampling.sample!

AdaptiveRejectionSampling.ARSampler — Type

ARSampler(
    obj::Objective{T, F, G},
    domain::Tuple{T, T},
    initial_points::Vector{T}
) where {T <: AbstractFloat, F <: Function, G <: Function}

ARSampler(
    obj::Objective{T, F, G},
    domain::Tuple{T, T},
    search_range::Tuple{T, T} = (-10.0, 10.0),
    search_step = 0.1
) where {T <: AbstractFloat, F <: Function, G <: Function}

Initialize an adaptive rejection sampler over an objective function from obj (AdaptiveRejectionSampling.Objective). initial_points should be a vector defining the abscissae of the initial segments of the sampler. At least 2 of the points should be on opposite sides of the objective function's maximum.

If no initial_points are provided an attempt to find suitable ones will be made by searching for the first negative/positive slope of obj. The default search range is -10:10 with a step of 0.1.

Warning

As it currently stands, finding inital points will fail for distributions bounded to the left/right of their maximum. In these cases the initial point(s) need to be provided manually (however, only a single initial point has to be provided).

AdaptiveRejectionSampling.Objective — Type

Objective(f::Function, init = one(Float64); adbackend = AutoForwardDiff())

Create an Objective for a function automatically generating its gradient. Gradients are calculated using DifferentiationInterface.jl using the backend of choice with ForwardDiff.jl being the default. In order to prepare the gradient an initial value is required. By default this is one(Float64). If a gradient for a different type is desired, it should be specified through this initial value.

If a custom/manual gradient is available it may instead be provided.

Warning

Observe that f should be in its log-concave form and that no checks are performed in order to verify this.

Example

# Create an objective using `Mooncake.jl` autodiff and Float32 as its type.
ARS.Objective(somefun, one(Float32); adbackend = AutoMooncake(; config=nothing))
# Create an objective using the defaults `AutoDiff.jl` autodiff and Float64 as its type.
ARS.Objective(somefun)

AdaptiveRejectionSampling.Objective — Method

Objective(f::Function, grad::Function, ::Type{T} = Float64) where {T <: Number}

Create an Objective directly defined by its function f and custom gradient grad. The parameter T represents the type of the expected input and is utilized when preparing gradients using autodiff. By default T = Float64.

Warning

Observe that f should be in its log-concave form and that no checks are performed in order to verify this.

AdaptiveRejectionSampling.sample! — Function

sample!([rng=default_rng()], s::ARSampler, n::Integer; add_segments::Bool=true, max_segments::Integer = DEFAULT_MAX_SEGMENTS)
sample!([rng=default_rng()], v::AbstractVector, s::ARSampler; add_segments::Bool=true, max_segments::Integer = DEFAULT_MAX_SEGMENTS)

Draw samples from s. If supplied, a vector v will be filled with samples. Otherwise the number of samples is specified with n.